Added new code for my home page
Added a style sheet in my head then put a nav bar in my bodies header after that a image of the logo in the header. after that made four paragraphs of the intro.next, made a hyperlink of the email and the phone number and finally added a footer.